backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-dotenv==1.0.0
httpx==0.24.1
beautifulsoup4==4.12.2
pandas==2.1.0
fredapi==0.5.0
python-dateutil==2.8.2
pydantic==2.3.0
requests==2.31.0
